Created folder for the sidebar
